Update primary routes to MVC structure
